feat(prover): update l1Current cursor to record L1 hash

+++ pkg/rpc/utils_test.go
@@ -40,7 +40,7 @@ func TestL1ContentFrom(t *testing.T) {
 	l2Head, err := client.L2.HeaderByNumber(context.Background(), nil)
 	require.Nil(t, err)
 
-	baseFee, err := client.TaikoL2.GetBasefee(nil, 0, 60000000, uint32(l2Head.GasUsed))
+	baseFee, err := client.TaikoL2.GetBasefee(nil, 0, 60000000, (l2Head.GasUsed))
 	require.Nil(t, err)
 
 	testAddrPrivKey, err := crypto.ToECDSA(common.Hex2Bytes(os.Getenv("L1_PROPOSER_PRIVATE_KEY")))
